Move md5_files_b64 to util
